LLM Projects/Code Converter/requirements.txt
python-dotenv
google-generativeai
gradio